+++ README.md
@@ -26,16 +26,15 @@ yarn add @floydnoel/react-kit
 import { Markdown } from '@floydnoel/react-kit';
 
 // ...
-
 // in a component
 <Markdown># hello, world</Markdown>;
 ```
 
 ## Components
 
-- Markdown [ [repo](/src/components/markdown), [docs](/src/components/markdown) ]
-- TextField [ [repo](https://github.com/floydnoel/react-kit/tree/babel/src/components/interaction/text-field), [docs](https://floydnoel.github.io/react-kit/text-field) ]
-- Json [ [repo](https://github.com/floydnoel/react-kit), [docs](https://floydnoel.github.io/react-kit/) ]
+- [Markdown](/src/components/markdown)
+- [TextField](/src/components/text-field)
+- [Json](/src/components/json)
 
 ## Core Ideals
